style: move custom color picker to main menu
